#2379 - Push recommender errors via websockets to user
- Folded LoggedEventMessageService into LoggedEventMessageController
